move response files and exceptions in the folders IntegradoResponses and IntegradoExceptions
