+++ TransbankPosSDK/CommonResponses/BasicResponse.cs
@@ -1,7 +1,7 @@
 ﻿using Transbank.POS.Utils;
 using System.Collections.Generic;
 
-namespace Transbank.POS.Responses
+namespace Transbank.POS.CommonResponses
 {
     public class BasicResponse
     {
