+++ TransbankPosSDK/CommonResponses/LoadKeysResponse.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System;
 
-namespace Transbank.POS.Responses
+namespace Transbank.POS.CommonResponses
 {
     public class LoadKeysResponse : BasicResponse
     {
