--- TransbankPosSDK/Exceptions/TransbankLoadKeysException.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using Transbank.POS.Responses;
-
-namespace Transbank.POS.Exceptions
-{
-    public class TransbankLoadKeysException : TransbankException
-    {
-        public LoadKeysResponse LoadKeyResponse;
-
-        public TransbankLoadKeysException(string message, LoadKeysResponse response) : base(message)
-        {
-            LoadKeyResponse = response;
-        }
-
-        public TransbankLoadKeysException(string message, Exception inner) : base(message, inner) { }
-    }
-}
