+++ TransbankPosSDK/IntegradoExceptions/TransbankLastSaleException.cs
@@ -1,9 +1,9 @@
 ﻿using System;
-using Transbank.POS.Responses;
+using Transbank.POS.IntegradoResponses;
 
-namespace Transbank.POS.Exceptions
+namespace Transbank.POS.IntegradoExceptions
 {
-    public class TransbankLastSaleException : TransbankException
+    public class TransbankLastSaleException : CommonExceptions.TransbankException
     {
         public LastSaleResponse LastSaleResponse;
