+++ TransbankPosSDK/IntegradoResponses/CloseResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace Transbank.POS.IntegradoResponses
+{
+    public class CloseResponse : CommonResponses.LoadKeysResponse
+    {
+        public CloseResponse(string response) : base(response) { }
+    }
+}
