--- TransbankPosSDK/Responses/CloseResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace Transbank.POS.Responses
-{
-    public class CloseResponse : LoadKeysResponse
-    {
-        public CloseResponse(string response) : base(response) { }
-    }
-}
