--- TransbankPosSDK/Responses/IntermediateResponse.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Transbank.POS.Responses
-{
-    public class IntermediateResponse : EventArgs
-    {
-        private readonly BasicResponse message;
-
-        public int FunctionCode => message.FunctionCode;
-        public string ResponseMessage => message.ResponseMessage;
-        public int ResponseCode => message.ResponseCode;
-
-        public IntermediateResponse(string response) => message = new BasicResponse(response);
-    }
-}
